refactor: rename app to wildthoughts
