+++ .gitignore
@@ -159,7 +159,7 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
-# animal_of_all_time stuff:
+# wildthoughts stuff:
 media/**/*
 !media/**/
 animal.json
