+++ templates/wildthoughts/base/base.html
@@ -18,7 +18,7 @@
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
             <ul class="navbar-nav me-auto mb-2 mb-lg-0">
                 <li class="nav-item">
-                    <a class="nav-link" aria-current="page" href="{% url 'animals-of-all-time:index' %}">Home</a></li>
+                    <a class="nav-link" aria-current="page" href="{% url 'wildthoughts:index' %}">Home</a></li>
                 <li class="nav-item dropdown">
                     <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Animals
@@ -26,16 +26,16 @@
                     <ul class="dropdown-menu">
                         <li><a class="dropdown-item" href="#">Most Overrated</a></li>
                         <li><a class="dropdown-item" href="#">Most Underrated</a></li>
-                        <li><a class="dropdown-item" href="{% url 'animals-of-all-time:animals' %}">Browse</a></li>
+                        <li><a class="dropdown-item" href="{% url 'wildthoughts:animals' %}">Browse</a></li>
                     </ul>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" aria-current="page" href="#">Lists</a></li>
                 <li class="nav-item">
-                    <a class="nav-link" aria-current="page" href="{% url 'animals-of-all-time:profiles' %}">Profiles</a>
+                    <a class="nav-link" aria-current="page" href="{% url 'wildthoughts:profiles' %}">Profiles</a>
                 </li>
             </ul>
-            <form class="d-flex" role="search" method=POST action="{% url 'animals-of-all-time:search' %}">
+            <form class="d-flex" role="search" method=POST action="{% url 'wildthoughts:search' %}">
                 {% csrf_token %}
                 <select class="form-control" name="category">
                     <option value="Animals" selected>Animals</option>
@@ -69,7 +69,7 @@
     {% block footer_block %}
     <footer class="py-2 mt-5">
         <div class="container text-center">
-            <p class="display-6 mb-2 text-black">Animals of All Times</p>
+            <p class="display-6 mb-2 text-black">WildThoughts</p>
             <small class="text-black-50">&copy; Copyright by Team 9A. All rights reserved.</small>
         </div>
     </footer>
