+++ templates/wildthoughts/profile/list_profiles.html
@@ -1,4 +1,4 @@
-{% extends 'animals_of_all_time/base/base.html' %}
+{% extends 'wildthoughts/base/base.html' %}
 {% load staticfiles %}
 
 {% block title_block %}
