+++ wad2_project/urls.py
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from animals_of_all_time import views
+from wildthoughts import views
 from django.conf.urls.static import static
 from django.conf import settings
 
@@ -23,5 +23,5 @@ from django.conf import settings
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', views.IndexView.as_view(), name='index'),
-    path('animals-of-all-time/', include('animals_of_all_time.urls')),
+    path('wildthoughts/', include('wildthoughts.urls')),
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
