+++ wildthoughts/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class AnimalsOfAllTimeConfig(AppConfig):
-    name = 'animals_of_all_time'
+    name = 'wildthoughts'
